Split customData field name handling, updated OrderBy
