+++ src/Stormpath.SDK.Core/Impl/Linq/Parsing/CompilingExpressionVisitor.cs
@@ -84,7 +84,14 @@ namespace Stormpath.SDK.Impl.Linq.Parsing
         internal Expression VisitOrderBy(OrderByExpression node)
         {
             string translatedFieldName = null;
-            if (!this.fieldNameTranslator.TryGetValue(node.FieldName, out translatedFieldName))
+
+            // Custom Data access doesn't need to be translated - use verbatim
+            if (node.FieldName.StartsWith("customData.", StringComparison.Ordinal))
+            {
+                translatedFieldName = node.FieldName;
+            }
+            // Other field names do (mostly for correct camelCasing; see the translator classes)
+            else if (!this.fieldNameTranslator.TryGetValue(node.FieldName, out translatedFieldName))
             {
                 throw new NotSupportedException($"{node.FieldName} is not a supported field.");
             }
@@ -135,7 +142,6 @@ namespace Stormpath.SDK.Impl.Linq.Parsing
             {
                 translatedFieldName = node.FieldName;
             }
-            
             // Other field names do (mostly for correct camelCasing; see the translator classes)
             else if (!this.fieldNameTranslator.TryGetValue(node.FieldName, out translatedFieldName) &&
                 !this.datetimeFieldNameTranslator.TryGetValue(node.FieldName, out translatedFieldName))
